+++ src/core/contexts.ts
@@ -2,43 +2,19 @@ import { createContext } from 'react'
 import type { Context } from 'react'
 
 import type { AnyAtom, Scope } from './types'
-import { State, UpdateState, createState } from './vanilla'
+import { UpdateState, createState } from './vanilla'
 import { createMutableSource } from './useMutableSource'
 
 type MutableSource = ReturnType<typeof createMutableSource>
 
-export type Store = [
-  mutableSource: MutableSource,
-  updateState: UpdateState,
-  getState: () => State
-]
+export type Store = [mutableSource: MutableSource, updateState: UpdateState]
 
 export const createStore = (
   initialValues?: Iterable<readonly [AnyAtom, unknown]>
 ): Store => {
-  let state = createState(initialValues)
-  type Updater = Parameters<UpdateState>[0]
-  const queue: Updater[] = []
-  const updateState = (updater: Updater) => {
-    queue.push(updater)
-    if (queue.length > 1) {
-      return
-    }
-    let nextState = state
-    while (queue.length) {
-      nextState = queue[0](nextState)
-      queue.shift()
-    }
-    if (nextState !== state) {
-      state = nextState
-      state.m.forEach((mounted) => {
-        mounted.l.forEach((listener) => listener())
-      })
-    }
-  }
-  const getState = () => state
-  const mutableSource = createMutableSource({ s: getState }, () => state)
-  return [mutableSource, updateState, getState]
+  const state = createState(initialValues)
+  const mutableSource = createMutableSource(state, () => state.v)
+  return [mutableSource, state.u]
 }
 
 type StoreContext = Context<ReturnType<typeof createStore>>
