+++ src/core/useAtom.ts
@@ -2,8 +2,8 @@ import { useContext, useCallback, useDebugValue } from 'react'
 
 import { getStoreContext } from './contexts'
 import { readAtom, writeAtom, subscribeAtom } from './vanilla'
-import type { State, UpdateState } from './vanilla'
-import type { Atom, WritableAtom, AnyWritableAtom, SetAtom } from './types'
+import type { State } from './vanilla'
+import type { Atom, WritableAtom, SetAtom } from './types'
 import { useMutableSource } from './useMutableSource'
 
 const isWritable = <Value, Update>(
@@ -20,8 +20,11 @@ export function useAtom<Value>(atom: Atom<Value>): [Value, never]
 export function useAtom<Value, Update>(
   atom: Atom<Value> | WritableAtom<Value, Update>
 ) {
+  const StoreContext = getStoreContext(atom.scope)
+  const [mutableSource, updateState] = useContext(StoreContext)
+
   const getAtomValue = useCallback(
-    (state: State, updateState: UpdateState) => {
+    (state: State) => {
       const atomState = readAtom(state, updateState, atom)
       if (atomState.e) {
         throw atomState.e // read error
@@ -37,34 +40,21 @@ export function useAtom<Value, Update>(
       }
       throw new Error('no atom value')
     },
-    [atom]
-  )
-
-  const StoreContext = getStoreContext(atom.scope)
-  const [mutableSource, updateState] = useContext(StoreContext)
-
-  const subscribe = useCallback(
-    (store: any, callback: () => void) =>
-      subscribeAtom(store.s(), updateState, atom, callback),
     [updateState, atom]
   )
 
-  const value: Value = useMutableSource(
-    mutableSource,
-    useCallback(
-      (store: any) => {
-        const state = store.s()
-        return getAtomValue(state, updateState)
-      },
-      [getAtomValue, updateState]
-    ),
-    subscribe
+  const subscribe = useCallback(
+    (state: State, callback: () => void) =>
+      subscribeAtom(state, updateState, atom, callback),
+    [updateState, atom]
   )
+
+  const value: Value = useMutableSource(mutableSource, getAtomValue, subscribe)
 
   const setAtom = useCallback(
     (update: Update) => {
       if (isWritable(atom)) {
-        return writeAtom(updateState, atom as AnyWritableAtom, update)
+        return writeAtom(updateState, atom, update)
       } else {
         throw new Error('not writable atom')
       }
